registered color theme to tailwind config
