+++ tailwind.config.js
@@ -1,7 +1,13 @@
 module.exports = {
   content: ['index.html', './src/**/*.{js,jsx,ts,tsx}'],
   theme: {
-    extend: {}
+    extend: {
+      colors: {
+        primary: '#073C64',
+        secondary: '#FDCB2C',
+        accent: '#1A2C43'
+      }
+    }
   },
   plugins: []
 }
